Firebase now works!

+++ GroceryAppV1/app/src/main/java/com/softwaredev/groceryappv1/PantryUI.java
@@ -68,12 +68,12 @@ public class PantryUI extends AppCompatActivity {
     Context context;
     static SharedPreferences sharedPref;
     int mSize;
-    boolean isPantry;
+    static boolean isPantry;
     private DrawerLayout mDrawerLayout;
     private ActionBarDrawerToggle mToggle;
     static String mUsername;
     static boolean isSignedIn = false;
-    User mUser;
+    static User mUser;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -153,6 +153,12 @@ public class PantryUI extends AppCompatActivity {
         mUsername = sharedPref.getString("username", "");
         isSignedIn = sharedPref.getBoolean("signedIn", false);
 
+        if (isPantry)
+            sharedPref = this.getSharedPreferences("com.softwaredev.groceryappv1.pantry", Context.MODE_PRIVATE);
+        else {
+            sharedPref = this.getSharedPreferences("com.softwaredev.groceryappv1.grocery", Context.MODE_PRIVATE);
+        }
+
         LoadList();
 
         FloatingActionButton fab = findViewById(R.id.switchToInput);
@@ -226,8 +232,13 @@ public class PantryUI extends AppCompatActivity {
                 return true;
             case R.id.removeAll:
                 pantry.clear();
-                editor.clear();
-                editor.commit();
+
+                if (!isSignedIn) {
+                    editor.clear();
+                    editor.commit();
+                }
+                else
+                    StoreInFirebase();
                 recreate();
                 return true;
             case R.id.addOne:
@@ -271,6 +282,12 @@ public class PantryUI extends AppCompatActivity {
 
         mUsername = sharedPref.getString("username", "");
         isSignedIn = sharedPref.getBoolean("signedIn", false);
+
+        if (isPantry)
+            sharedPref = this.getSharedPreferences("com.softwaredev.groceryappv1.pantry", Context.MODE_PRIVATE);
+        else {
+            sharedPref = this.getSharedPreferences("com.softwaredev.groceryappv1.grocery", Context.MODE_PRIVATE);
+        }
 
 
         LoadList();
@@ -332,19 +349,13 @@ public class PantryUI extends AppCompatActivity {
         pantry.add(_item);
         if (!isSignedIn) {
             SharedPreferences.Editor editor = sharedPref.edit();
-            editor.putString("item" + Integer.toString(pantry.size()), _item.itemToString());
+            editor.putString("item" + Integer.toString(pantry.size() - 1), _item.itemToString());
 
             editor.putInt("size", pantry.size());
             editor.commit();
         }
         else {
-            mRef = FirebaseDatabase.getInstance().getReference();
-            User mUser = new User();
-            mUser.setPantry(pantry);
-            mUser.setPantrySize(pantry.size());
-            mUser.setUsername(mUsername);
-            mRef.child(mUsername).setValue(mUser);
-            //mRef.child("pantry").child("size").setValue(pantry.size());
+            StoreInFirebase();
         }
     }
 
@@ -368,9 +379,15 @@ public class PantryUI extends AppCompatActivity {
         pantry.get(position).mYear = year;
         pantry.get(position).mQuantity = quantity;
 
-        SharedPreferences.Editor editor = sharedPref.edit();
-        editor.putString("item" + Integer.toString(position), pantry.get(position).itemToString());
-        editor.commit();
+        if (!isSignedIn) {
+            SharedPreferences.Editor editor = sharedPref.edit();
+            editor.putString("item" + Integer.toString(position), pantry.get(position).itemToString());
+            editor.commit();
+        }
+        else
+        {
+            StoreInFirebase();
+        }
     }
 
     public static void addToQuantity(int position, int quantity)
@@ -382,9 +399,13 @@ public class PantryUI extends AppCompatActivity {
         else
             pantry.get(position).mQuantity = 999999999;
 
-        SharedPreferences.Editor editor = sharedPref.edit();
-        editor.putString("item" + Integer.toString(position), pantry.get(position).itemToString());
-        editor.commit();
+        if (!isSignedIn) {
+            SharedPreferences.Editor editor = sharedPref.edit();
+            editor.putString("item" + Integer.toString(position), pantry.get(position).itemToString());
+            editor.commit();
+        }
+        else
+            StoreInFirebase();
     }
 
     public static void removeFromList(int position)
@@ -392,16 +413,21 @@ public class PantryUI extends AppCompatActivity {
         if (position > -1) {
             pantry.remove(position);
 
-            SharedPreferences.Editor editor = sharedPref.edit();
-            editor.clear();
-            editor.commit();
+            if (!isSignedIn) {
+                SharedPreferences.Editor editor = sharedPref.edit();
+                editor.clear();
+                editor.commit();
 
-            editor.putInt("size", pantry.size());
-            for (int i = 0; i < pantry.size(); ++i)
-            {
-                editor.putString("item" + Integer.toString(i), pantry.get(i).itemToString());
+                editor.putInt("size", pantry.size());
+                for (int i = 0; i < pantry.size(); ++i) {
+                    editor.putString("item" + Integer.toString(i), pantry.get(i).itemToString());
+                }
+                editor.commit();
             }
-            editor.commit();
+            else{
+                StoreInFirebase();
+            }
+
         }
     }
 
@@ -564,6 +590,7 @@ public class PantryUI extends AppCompatActivity {
 
     public void LoadList () {
         if (!isSignedIn) {
+
             if (isPantry)
                 sharedPref = this.getSharedPreferences("com.softwaredev.groceryappv1.pantry", Context.MODE_PRIVATE);
             else {
@@ -632,6 +659,21 @@ public class PantryUI extends AppCompatActivity {
                 totalText.setText("Total: $" + total);
             }
 
+    }
+
+    public static void StoreInFirebase () {
+        mRef = FirebaseDatabase.getInstance().getReference();
+
+        if (isPantry) {
+            mUser.setPantry(pantry);
+            mUser.setPantrySize(pantry.size());
+        }
+        else {
+            mUser.setGrocery(pantry);
+            mUser.setGrocerySize(pantry.size());
+        }
+        mUser.setUsername(mUsername);
+        mRef.child(mUsername).setValue(mUser);
     }
 }
